+++ PlantUmlViewer/PlantUmlViewer/Forms/OptionsWindow.Designer.cs
@@ -1,6 +1,6 @@
 ﻿namespace PlantUmlViewer.Forms
 {
-    partial class SettingsWindow
+    partial class OptionsWindow
     {
         /// <summary>
         /// Required designer variable.
@@ -30,12 +30,12 @@
         {
             this.tableLayoutPanel_Window = new System.Windows.Forms.TableLayoutPanel();
             this.label_headerGeneral = new System.Windows.Forms.Label();
+            this.tableLayoutPanel_Buttons = new System.Windows.Forms.TableLayoutPanel();
+            this.button_Cancel = new System.Windows.Forms.Button();
+            this.button_ok = new System.Windows.Forms.Button();
             this.label_JavaPath = new System.Windows.Forms.Label();
             this.textBox_JavaPath = new System.Windows.Forms.TextBox();
             this.label_JavaPathDescription = new System.Windows.Forms.Label();
-            this.button_ok = new System.Windows.Forms.Button();
-            this.button_Cancel = new System.Windows.Forms.Button();
-            this.tableLayoutPanel_Buttons = new System.Windows.Forms.TableLayoutPanel();
             this.tableLayoutPanel_Window.SuspendLayout();
             this.tableLayoutPanel_Buttons.SuspendLayout();
             this.SuspendLayout();
@@ -75,6 +75,45 @@
             this.label_headerGeneral.TabIndex = 3;
             this.label_headerGeneral.Text = "General";
             // 
+            // tableLayoutPanel_Buttons
+            // 
+            this.tableLayoutPanel_Buttons.ColumnCount = 2;
+            this.tableLayoutPanel_Window.SetColumnSpan(this.tableLayoutPanel_Buttons, 2);
+            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 50F));
+            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 50F));
+            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this.tableLayoutPanel_Buttons.Controls.Add(this.button_Cancel, 0, 0);
+            this.tableLayoutPanel_Buttons.Controls.Add(this.button_ok, 1, 0);
+            this.tableLayoutPanel_Buttons.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.tableLayoutPanel_Buttons.Location = new System.Drawing.Point(8, 232);
+            this.tableLayoutPanel_Buttons.Name = "tableLayoutPanel_Buttons";
+            this.tableLayoutPanel_Buttons.RowCount = 1;
+            this.tableLayoutPanel_Buttons.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
+            this.tableLayoutPanel_Buttons.Size = new System.Drawing.Size(550, 34);
+            this.tableLayoutPanel_Buttons.TabIndex = 2;
+            // 
+            // button_Cancel
+            // 
+            this.button_Cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
+            this.button_Cancel.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.button_Cancel.Location = new System.Drawing.Point(3, 3);
+            this.button_Cancel.Name = "button_Cancel";
+            this.button_Cancel.Size = new System.Drawing.Size(269, 28);
+            this.button_Cancel.TabIndex = 1;
+            this.button_Cancel.Text = "Cancel";
+            this.button_Cancel.UseVisualStyleBackColor = true;
+            // 
+            // button_ok
+            // 
+            this.button_ok.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.button_ok.Location = new System.Drawing.Point(278, 3);
+            this.button_ok.Name = "button_ok";
+            this.button_ok.Size = new System.Drawing.Size(269, 28);
+            this.button_ok.TabIndex = 0;
+            this.button_ok.Text = "OK";
+            this.button_ok.UseVisualStyleBackColor = true;
+            this.button_ok.Click += new System.EventHandler(this.Button_Ok_Click);
+            // 
             // label_JavaPath
             // 
             this.label_JavaPath.AutoSize = true;
@@ -107,46 +146,7 @@
     " to auto detect.";
             this.label_JavaPathDescription.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
             // 
-            // button_ok
-            // 
-            this.button_ok.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.button_ok.Location = new System.Drawing.Point(278, 3);
-            this.button_ok.Name = "button_ok";
-            this.button_ok.Size = new System.Drawing.Size(269, 28);
-            this.button_ok.TabIndex = 0;
-            this.button_ok.Text = "OK";
-            this.button_ok.UseVisualStyleBackColor = true;
-            this.button_ok.Click += new System.EventHandler(this.Button_Ok_Click);
-            // 
-            // button_Cancel
-            // 
-            this.button_Cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.button_Cancel.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.button_Cancel.Location = new System.Drawing.Point(3, 3);
-            this.button_Cancel.Name = "button_Cancel";
-            this.button_Cancel.Size = new System.Drawing.Size(269, 28);
-            this.button_Cancel.TabIndex = 1;
-            this.button_Cancel.Text = "Cancel";
-            this.button_Cancel.UseVisualStyleBackColor = true;
-            // 
-            // tableLayoutPanel_Buttons
-            // 
-            this.tableLayoutPanel_Buttons.ColumnCount = 2;
-            this.tableLayoutPanel_Window.SetColumnSpan(this.tableLayoutPanel_Buttons, 2);
-            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 50F));
-            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Percent, 50F));
-            this.tableLayoutPanel_Buttons.ColumnStyles.Add(new System.Windows.Forms.ColumnStyle(System.Windows.Forms.SizeType.Absolute, 20F));
-            this.tableLayoutPanel_Buttons.Controls.Add(this.button_Cancel, 0, 0);
-            this.tableLayoutPanel_Buttons.Controls.Add(this.button_ok, 1, 0);
-            this.tableLayoutPanel_Buttons.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.tableLayoutPanel_Buttons.Location = new System.Drawing.Point(8, 232);
-            this.tableLayoutPanel_Buttons.Name = "tableLayoutPanel_Buttons";
-            this.tableLayoutPanel_Buttons.RowCount = 1;
-            this.tableLayoutPanel_Buttons.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 100F));
-            this.tableLayoutPanel_Buttons.Size = new System.Drawing.Size(550, 34);
-            this.tableLayoutPanel_Buttons.TabIndex = 2;
-            // 
-            // SettingsWindow
+            // OptionsWindow
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
@@ -156,11 +156,11 @@
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
             this.MaximizeBox = false;
             this.MinimizeBox = false;
-            this.Name = "SettingsWindow";
+            this.Name = "OptionsWindow";
             this.ShowIcon = false;
             this.ShowInTaskbar = false;
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
-            this.Text = "Settings";
+            this.Text = "Options";
             this.tableLayoutPanel_Window.ResumeLayout(false);
             this.tableLayoutPanel_Window.PerformLayout();
             this.tableLayoutPanel_Buttons.ResumeLayout(false);