+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # X.X.X.X (XX.XX.XXXX)
 - Added support for dark interfaces
+- Extended icon to highlight itself and close preview if opened
+- Added menu entry and shortcut Shift + U to refresh diagram
+- Renamed settings to options
 
 # 1.0.2.3 (22.05.2022)
 - Fixed generation fails for release builds